Move SMBServer into its own file
